+++ core/model/member.go
@@ -14,7 +14,10 @@
 
 package model
 
-import "time"
+import (
+	"groups/driven/notifications"
+	"time"
+)
 
 // Member represents group member entity
 type Member struct {
@@ -62,6 +65,24 @@ func (m *Member) ApplyFromUserIfEmpty(user *User) {
 	}
 }
 
+// ToMember construct ToMember based on the data
+func (m *Member) ToMember() ToMember {
+	return ToMember{
+		UserID:     m.UserID,
+		ExternalID: m.ExternalID,
+		Name:       m.Name,
+		Email:      m.Email,
+	}
+}
+
+// ToNotificationRecipient construct notifications.Recipient based on the data
+func (m *Member) ToNotificationRecipient() notifications.Recipient {
+	return notifications.Recipient{
+		UserID: m.UserID,
+		Name:   m.Name,
+	}
+}
+
 // ToMember represents to(destination) member entity
 type ToMember struct {
 	UserID     string `json:"user_id" bson:"user_id"`
@@ -100,3 +121,27 @@ func (m *Member) IsPendingMember() bool {
 func (m *Member) IsRejected() bool {
 	return m.Status == "rejected"
 }
+
+//ToShortMemberRecord converts to ShortMemberRecord
+func (m *Member) ToShortMemberRecord() ShortMemberRecord {
+	return ShortMemberRecord{
+		ID:         m.ID,
+		UserID:     m.UserID,
+		ExternalID: m.ExternalID,
+		Email:      m.Email,
+		NetID:      m.NetID,
+		Name:       m.Name,
+		Status:     m.Status,
+	}
+}
+
+// ShortMemberRecord represents group short member entity only with important identifiers
+type ShortMemberRecord struct {
+	ID         string `json:"id" bson:"id"`
+	UserID     string `json:"user_id" bson:"user_id"`
+	ExternalID string `json:"external_id" bson:"external_id"`
+	Name       string `json:"name" bson:"name"`
+	NetID      string `json:"net_id" bson:"net_id"`
+	Email      string `json:"email" bson:"email"`
+	Status     string `json:"status" bson:"status"` //pending, member, admin, rejected
+} //@name ShortMemberRecord